+++ src/main/java/com/influxdb/v3/client/internal/InfluxDBClientImpl.java
@@ -398,13 +398,6 @@ public final class InfluxDBClientImpl implements InfluxDBClient {
         Arguments.checkNotNull(parameters, "parameters");
         Arguments.checkNotNull(options, "options");
 
-        if (options.grpcCallOptions().getDeadline() == null && config.getQueryTimeout() != null) {
-            options.setGrpcCallOptions(new GrpcCallOptions.Builder()
-                .fromGrpcCallOptions(options.grpcCallOptions())
-                .withDeadline(Deadline.after(config.getQueryTimeout().toMillis(), TimeUnit.MILLISECONDS))
-                .build());
-        }
-
         if (closed) {
             throw new IllegalStateException("InfluxDBClient has been closed.");
         }
@@ -421,7 +414,31 @@ public final class InfluxDBClientImpl implements InfluxDBClient {
             }
         });
 
-        CallOption[] callOptions = options.grpcCallOptions().getCallOptions();
+        GrpcCallOptions.Builder builder = new GrpcCallOptions.Builder()
+            .fromGrpcCallOptions(options.grpcCallOptions());
+
+        if (config.getQueryTimeout() == null) {
+            if (options.grpcCallOptions().getDeadline() != null
+                && options.grpcCallOptions().getDeadline().timeRemaining(TimeUnit.MILLISECONDS) <= 0) {
+                LOG.warning("Query timeout "
+                    + options.grpcCallOptions().getDeadline()
+                    + " is 0 or negative and will be ignored.");
+                builder.withoutDeadline();
+            }
+        } else {
+            if (options.grpcCallOptions().getDeadline() == null) {
+                builder.withDeadline(Deadline.after(config.getQueryTimeout().toMillis(), TimeUnit.MILLISECONDS));
+            } else if (options.grpcCallOptions().getDeadline().timeRemaining(TimeUnit.MILLISECONDS) <= 0) {
+                LOG.warning("Query timeout "
+                    +  options.grpcCallOptions().getDeadline()
+                    + " is 0 or negative. Using config.queryTimeout "
+                    + config.getQueryTimeout()
+                    + " instead.");
+                builder.withDeadline(Deadline.after(config.getQueryTimeout().toMillis(), TimeUnit.MILLISECONDS));
+            }
+        }
+
+        CallOption[] callOptions = builder.build().getCallOptions();
 
         return flightSqlClient.execute(
                 query,